long values to double, commentary in Process

+++ SO-simulation_csharp/FCFS.cs
@@ -27,7 +27,7 @@ namespace SO_simulation_csharp
         /// <summary>
         /// Zmienna wykorzystywana do mierzenia taktow zegarowych
         /// </summary>
-        private long cyclesNumber;
+        private double cyclesNumber;
 
         /// <summary>
         /// Konstruktor wczytujacy ciagi procesow z zadanej sciezki, przechowywane w odpowiednim formacie XML
@@ -63,10 +63,10 @@ namespace SO_simulation_csharp
         /// Funkcja zwraca liste srednich czasow oczekiwania dla poszczegolnych ciagow procesow
         /// </summary>
         /// <returns>Lista srednich czasow oczekiwania</returns>
-        public List<long> AverageWaitingTimeForEachSequence()
+        public List<double> AverageWaitingTimeForEachSequence()
         {
-            List<long> listOfWaitingTimes = new List<long>();
-            long waitingTime = 0;
+            List<double> listOfWaitingTimes = new List<double>();
+            double waitingTime = 0;
             foreach (List<Process> list in LoadedProcesses)
             {
                 foreach (Process process in list)
@@ -83,10 +83,10 @@ namespace SO_simulation_csharp
         /// Funkcja zwraca liste srednich czasow przetwarzania dla poszczegolnych ciagow procesow
         /// </summary>
         /// <returns>Lista srednich czasow oczekiwania</returns>
-        public List<long> AverageTurnaroundTimeForEachSequence()
+        public List<double> AverageTurnaroundTimeForEachSequence()
         {
-            List<long> listOfTurnaroundTime = new List<long>();
-            long turnaroundTime = 0;
+            List<double> listOfTurnaroundTime = new List<double>();
+            double turnaroundTime = 0;
 
             foreach (List<Process> list in LoadedProcesses)
             {
@@ -106,8 +106,8 @@ namespace SO_simulation_csharp
         /// </summary>
         public void PrintFCFSResults()
         {
-            long averageWaitingTime = 0;
-            long averageTurnaroundTime = 0;
+            double averageWaitingTime = 0;
+            double averageTurnaroundTime = 0;
             for (int i = 0; i < processUtilities.AmountOfProcessesLists; i++)
             {
                 averageWaitingTime += AverageWaitingTimeForEachSequence().ElementAt(i);
